text2sql_fastapi/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
sqlalchemy==2.0.29
pydantic==2.9.2
pandas==2.2.3
polars==1.11.0
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.9
langchain-experimental==0.3.2
langchain-openai==0.2.2
langchain-ollama==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.129
requests==2.31.0
PyMuPDF==1.24.11
Pillow==10.3.0
pymysql==1.1.1